Move configs under advanced tab for preload to work
