+++ src/linodes/linode/settings/advanced/layouts/ConfigsDisksPage.js
@@ -1,33 +1,14 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
-import { setError } from '~/actions/errors';
 import { setSource } from '~/actions/source';
-import { linodes, kernels } from '~/api';
-import { getObjectByLabelLazily } from '~/api/util';
 
 import { ConfigPanel } from '../components/ConfigPanel';
 import { DiskPanel } from '../components/DiskPanel';
-import { selectLinode } from '../../utilities';
+import { selectLinode } from '../../../utilities';
 
 
 export class AdvancedPage extends Component {
-  static async preload({ dispatch, getState }, { linodeLabel }) {
-    const { id } = await dispatch(getObjectByLabelLazily('linodes', linodeLabel));
-
-    try {
-      await dispatch(linodes.disks.all([id]));
-      const { total_pages: totalPages } = await dispatch(kernels.page(0));
-
-      // Fetch rest of things without waiting.
-      for (let i = 1; i < totalPages; i++) {
-        dispatch(kernels.page(i));
-      }
-    } catch (e) {
-      dispatch(setError(e));
-    }
-  }
-
   componentDidMount() {
     const { dispatch } = this.props;
     dispatch(setSource(__filename));
