+++ src/linodes/linode/settings/advanced/layouts/index.js
@@ -1,0 +1,4 @@
+export ConfigsDisksPage from './ConfigsDisksPage';
+export EditConfigPage from './EditConfigPage';
+export AddConfigPage from './AddConfigPage';
+export IndexPage from './IndexPage';
